fix: structure updated
