+++ src/app/members/actions/createUser.ts
@@ -2,7 +2,7 @@
 import { Tables } from "@/lib/db";
 import { UserStatus } from "@/constant/constant";
 import { redirect } from "next/navigation";
-import { formatPhoneNumber } from "../../../../utils/formatPhoneNumber";
+import { intlNumberFormat } from "@/utils/phoneNumberFormatter";
 import { UserBuilder } from "../components/MemberFormBuilder";
 import { headers } from "next/headers";
 import { getSupabaseClient } from "@/utils/supabase/supabaseClient";
@@ -20,7 +20,7 @@ export const createUser = async (values: UserBuilder) => {
 
   const invite_link = `${protocol}://${host}?token=${randomInvitedId}`;
 
-  const userPhoneNumber = formatPhoneNumber(values.phoneNumber);
+  const userPhoneNumber = intlNumberFormat(values.phoneNumber);
 
   const { error } = await supabase.from(Tables.Members).insert({
     ...values,
